Parse questData PayEntityType into new OtherCostType field
As prepared in https://dragalialost.gamepedia.com/index.php?title=Template:QuestDisplay&diff=68216&oldid=67628, Quest pay entities that are not normal units like stamina or wings will be grouped similar to the game hierarchy as a single type-quantity pair in the output template, instead of individually having parameters. Updating the parsing to reflect this.